Hirejoure.com/requirements.txt
playwright>=1.52.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
loguru>=0.7.0
retrying>=1.3.4
aiohttp>=3.9.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
matplotlib>=3.7.2
seaborn>=0.12.2
wordcloud>=1.9.3
networkx>=3.1
numpy>=1.24.31
Django>=4.2
djangorestframework>=3.15
